requirements.txt
requests
beautifulsoup4
pandas
openpyxl
google-cloud-storage
deep-translator
langdetect

scikit-learn
matplotlib
seaborn
